subspace_runtime_primitives

Module extension

Structs

BalanceTransferCheckExtension

Disable balance transfers, if configured in the runtime.

Enums

Pre

Data passed from prepare to post_dispatch.

Val

Data passed from validate to prepare.

Traits

BalanceTransferChecks

Trait to check if the Balance transfers are enabled.

MaybeBalancesCall

Trait to convert Runtime call to possible Balance call.

WeightInfo

Weights for the balance transfer check extension.