Module subspace_runtime_primitives::time

Constants

BLOCKS_IN_AN_HOUR

Approximate number of blocks in an hour.

BLOCKS_IN_AN_MINUTE

Approximate number of block in a minute.

BLOCKS_IN_A_DAY

Approximate number of blocks in a day.

MILLISECS_PER_BLOCK

Since Subspace is probabilistic this is the average expected block time that we are targeting. Blocks will be produced at a minimum duration defined by SLOT_DURATION, but some slots will not be allocated to any farmer and hence no block will be produced. We expect to have this block time on average following the defined slot duration and the value of c configured for Subspace (where 1 - c represents the probability of a slot being empty). This value is only used indirectly to define the unit constants below that are expressed in blocks. The rest of the code should use SLOT_DURATION instead (like the Timestamp pallet for calculating the minimum period).