subspace_runtime_primitives::utility

Trait MaybeNestedCall

pub trait MaybeNestedCall<Runtime: Config> {
    // Required method
    fn maybe_nested_call(&self) -> Option<Vec<&RuntimeCallFor<Runtime>>>;
}

Trait used to extract nested RuntimeCalls from a RuntimeCall type. Each runtime has a different set of pallets which can nest calls.

Required Methods

fn maybe_nested_call(&self) -> Option<Vec<&RuntimeCallFor<Runtime>>>

If this call is a nested runtime call, returns the inner call(s).

Ignored calls (such as pallet_utility::Call::__Ignore) should be yielded themsevles, but their contents should not be yielded.